separated tests on vertices and edges into 2 files
