+++ tests/test_edges.py
@@ -3,21 +3,6 @@ from graphtool.graph import *
 from utils import *
 
 
-def test_vertex_eq():
-    vertex1 = Vertex(0)
-    assert vertex1 == 0
-
-
-def test_vertex_lt():
-    vertex1 = Vertex(0)
-    assert vertex1 < 3
-
-
-def test_vertex_init():
-    vertex1 = Vertex(0, data={})
-    assert vertex1["name"] == "0"
-
-
 def test_edge():
     edge1 = Edge(0, 1)
     edge2 = Edge(Vertex(0), Vertex(1))
